01-Python/Python-Projects/Growth-Mind-Set-App/requirements.txt
streamlit
opencv-python-headless
numpy
matplotlib
scikit-learn
webcolors
mysql-connector-python
